fixed conflict

+++ IslandFurniture/IslandFurniture-ejb/src/java/IslandFurniture/Entities/FurnitureTransactionDetail.java
@@ -28,7 +28,7 @@ public class FurnitureTransactionDetail implements Serializable {
     private Integer numReturned;
     private Integer numClaimed;
     private Double unitPrice;
-    private Integer unitPoints;
+    private Long unitPoints;
 
     @ManyToOne
     private FurnitureTransaction furnitureTransaction;
@@ -77,11 +77,11 @@ public class FurnitureTransactionDetail implements Serializable {
         this.unitPrice = unitPrice;
     }
 
-    public Integer getUnitPoints() {
+    public Long getUnitPoints() {
         return unitPoints;
     }
 
-    public void setUnitPoints(Integer unitPoints) {
+    public void setUnitPoints(Long unitPoints) {
         this.unitPoints = unitPoints;
     }
 
@@ -139,7 +139,7 @@ public class FurnitureTransactionDetail implements Serializable {
         return this.qty * this.unitPrice;
     }
 
-    public Integer getTotalPoints() {
+    public Long getTotalPoints() {
         return this.qty * this.unitPoints;
     }
